pyproject.toml
[project]
name = "brat-vertix"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-cloud-storage>=3.9.0",
    "mlflow>=3.10.0",
    "monai>=1.5.2",
    "nibabel>=5.3.3",
    "numpy>=2.2.6",
    "python-dotenv>=1.2.1",
    "python-json-logger>=4.0.0",
    "ruff>=0.15.4",
    "scikit-learn>=1.7.2",
    "torch>=2.10.0",
    "tqdm>=4.67.3",
]
